refactor(db-types): move the cashu/spark json-model schemas into @agicash/db-types
Now that ProofSchema lives in @agicash/cashu, the remaining agicash-db json-models (jsonb
column schemas) can move into @agicash/db-types where they belong:

- cashu-lightning-receive, cashu-swap-receive, cashu-swap-send, cashu-token-melt,
  spark-lightning-receive, spark-lightning-send -> packages/db-types/src/json-models
  (byte-identical git mv; ~/lib/money -> @agicash/utils/money, ~/lib/cashu ProofSchema ->
  @agicash/cashu)
- old paths kept as transitional re-export shims so the json-models barrel + its 15 consumers
  stay untouched
- db-types now depends on @agicash/cashu (ProofSchema) + @agicash/utils (Money) — this coupling
  already existed in the app; the package graph now just reflects it. db-types tsconfig gains the
  DOM lib (it transitively pulls in money.ts's devtools-formatter window use)

Staying in the app for their own phases: cashu-lightning-send-db-data (needs the send-domain
DestinationDetailsSchema) and account-details-db-data (account-domain combined schema).

Phase 0. typecheck, biome, build all green; 133 tests preserved.
